+++ ktlint-ruleset-standard/src/main/kotlin/com/pinterest/ktlint/ruleset/standard/StandardRuleSetProvider.kt
@@ -10,6 +10,7 @@ public class StandardRuleSetProvider : RuleSetProvider {
     override fun get(): RuleSet = RuleSet(
         "standard",
         AnnotationRule(),
+        AnnotationSpacingRule(),
         ChainWrappingRule(),
         CommentSpacingRule(),
         FilenameRule(),