+++ CHANGELOG.md
@@ -8,6 +8,7 @@ This project adheres to [Semantic Versioning](https://semver.org/).
 
 The rules below are promoted from the Experiment ruleset to the Standard ruleset.
 * `annotation`
+* `annotation-spacing`
 * `traling-comma`
 
 Note that as a result of moving the rules that the prefix `experimental:` has to be removed from all references to this rule. Check references in: